requirements.txt
ffmpeg-python==0.2.0
python-telegram-bot==12.8
PyWit==0.4.0
